demo/requirements.txt
django-ordered-model
django-bootstrap3
Sphinx
django-debug-toolbar
django-extensions<3.2.0  # The version requirement can be removed once https://github.com/django-extensions/django-extensions/issues/1743 is fixed
mock
coverage
tox
maxminddb-geolite2
model_bakery
django-concurrent-test-helper
freezegun
flake8
psycopg2<2.9
internet-sabotage3>=0.1.6
django_coverage_plugin
swapper


# Not supported in py3
#fabric
#transifex-client
